warning fix

+++ TTT/ViewControllers/GameBoardViewController.m
@@ -81,8 +81,8 @@
     [lbl_Player1 setText:_Player1Name];
     [lbl_Player2 setText:_Player2Name];
     
-    [lbl_score setText:[NSString stringWithFormat:@"%d (%d tie(s)) %d", Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), Player2Wins]];
-    [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%d game(s))", totalMatches]];
+    [lbl_score setText:[NSString stringWithFormat:@"%lu (%lu tie(s)) %lu", (unsigned long)Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), (unsigned long)Player2Wins]];
+    [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%lu game(s))", (unsigned long)totalMatches]];
 }
 
 -(void)resetGameBoard
@@ -299,8 +299,8 @@
             gameResultAlert = [[UIAlertView alloc] initWithTitle:@"Its a Tie" message:@"Try again?" delegate:self cancelButtonTitle:@"Cancel" otherButtonTitles:@"Retry", nil];
             [gameResultAlert show];
             ++totalMatches;
-            [lbl_score setText:[NSString stringWithFormat:@"%d (%d tie(s)) %d", Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), Player2Wins]];
-            [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%d game(s))", totalMatches]];
+            [lbl_score setText:[NSString stringWithFormat:@"%lu (%lu tie(s)) %lu", (unsigned long)Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), (unsigned long)Player2Wins]];
+            [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%lu game(s))", (unsigned long)totalMatches]];
             _gameOver = TRUE;
             return;
         }
@@ -314,8 +314,8 @@
                 //[gameResultAlert show];
                 ++totalMatches;
                 ++Player1Wins;
-                [lbl_score setText:[NSString stringWithFormat:@"%d (%d tie(s)) %d", Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), Player2Wins]];
-                [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%d game(s))", totalMatches]];
+                [lbl_score setText:[NSString stringWithFormat:@"%lu (%lu tie(s)) %lu", (unsigned long)Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), (unsigned long)Player2Wins]];
+                [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%lu game(s))", (unsigned long)totalMatches]];
                 [self animateWinner:Player_1_Turn];
                 _gameOver = TRUE;
                 return;
@@ -340,8 +340,8 @@
                 //[gameResultAlert show];
                 ++totalMatches;
                 ++Player2Wins;
-                [lbl_score setText:[NSString stringWithFormat:@"%d (%d tie(s)) %d", Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), Player2Wins]];
-                [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%d game(s))", totalMatches]];
+                [lbl_score setText:[NSString stringWithFormat:@"%lu (%lu tie(s)) %lu", (unsigned long)Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), (unsigned long)Player2Wins]];
+                [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%lu game(s))", (unsigned long)totalMatches]];
                 [self animateWinner:Player_2_Turn];
                 _gameOver = TRUE;
                 return;
@@ -789,8 +789,8 @@
         gameResultAlert = [[UIAlertView alloc] initWithTitle:@"Its a Tie" message:@"Try again?" delegate:self cancelButtonTitle:@"Cancel" otherButtonTitles:@"Retry", nil];
         [gameResultAlert show];
         ++totalMatches;
-        [lbl_score setText:[NSString stringWithFormat:@"%d (%d tie(s)) %d", Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), Player2Wins]];
-        [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%d game(s))", totalMatches]];
+        [lbl_score setText:[NSString stringWithFormat:@"%lu (%lu tie(s)) %lu", (unsigned long)Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), (unsigned long)Player2Wins]];
+        [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%lu game(s))", (unsigned long)totalMatches]];
         _gameOver = TRUE;
         return;
     }
@@ -804,8 +804,8 @@
         
         ++Player2Wins;
         ++totalMatches;
-        [lbl_score setText:[NSString stringWithFormat:@"%d (%d tie(s)) %d", Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), Player2Wins]];
-        [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%d game(s))", totalMatches]];
+        [lbl_score setText:[NSString stringWithFormat:@"%lu (%lu tie(s)) %lu", (unsigned long)Player1Wins,(totalMatches - (Player2Wins + Player1Wins)), (unsigned long)Player2Wins]];
+        [lbl_Totalmatches setText:[NSString stringWithFormat:@"(%lu game(s))", (unsigned long)totalMatches]];
         [self animateWinner:Player_2_Turn];
         _gameOver = TRUE;
         return;
